CogVM source as per VMMaker.oscog-eem.974
Fix 64-bit issues in the AsynchFilePlugin.
Fix 64-bit issues with sqUnixMain.c, sqMemoryFence.h & sqVirtualMachine.c.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3169 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: d483c2d70b3f660e79cb33c3f69097988018a226

+++ platforms/Cross/vm/sqVirtualMachine.c
@@ -556,7 +556,7 @@ pushOutputFile(char *filenameOrStdioIndex)
 		fprintf(stderr,"output file stack is full.\n");
 		return;
 	}
-	switch ((unsigned)filenameOrStdioIndex) {
+	switch ((unsigned long)filenameOrStdioIndex) {
 	case STDOUT_FILENO: output = stdout; break;
 	case STDERR_FILENO: output = stderr; break;
 	default:
